+++ src/include/tclap/UnlabeledMultiArg.h
@@ -143,7 +143,7 @@ public:
      * \param i - Pointer the the current argument in the list.
      * \param args - Mutable list of strings. Passed from main().
      */
-    virtual bool processArg(int *i, std::vector<std::string> &args);
+    virtual bool processArg(size_t *i, std::vector<std::string> &args);
 
     /**
      * Returns the a short id string.  Used in the usage.
@@ -221,7 +221,7 @@ UnlabeledMultiArg<T>::UnlabeledMultiArg(const std::string &name,
 }
 
 template <class T>
-bool UnlabeledMultiArg<T>::processArg(int *i, std::vector<std::string> &args) {
+bool UnlabeledMultiArg<T>::processArg(size_t *i, std::vector<std::string> &args) {
     if (_hasBlanks(args[*i])) return false;
 
     // never ignore an unlabeled multi arg
